// web/src/api-access/generated/models/schedulingTimeOffSeriesRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Unified.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface SchedulingTimeOffSeriesRequest {
  recurrenceRule: string;
  leaveTypeId?: number;
  userIds?: string[];
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  color?: string | null;
  startAtUtc: string;
  /** @nullable */
  endAtUtc?: string | null;
  /** @nullable */
  seriesStartAtUtc?: string | null;
  /** @nullable */
  seriesEndAtUtc?: string | null;
  /** @nullable */
  timeZoneId?: string | null;
  allDay?: boolean;
  locationId?: number;
}